가게 상세페이지 개발

+++ backend/src/main/java/com/myong/backend/domain/dto/shop/ShopDesignerDetailResponseDto.java
@@ -20,5 +20,5 @@ public class ShopDesignerDetailResponseDto {
     String gender;// 디자이너 성별
 
     @NotBlank
-    Long like; // 디자이너 좋아요 개수
+    Integer like; // 디자이너 좋아요 개수
 }
